Example MNIST code for machine learning-tutorial
This patch add a script for MNIST example with tenserflow. The example
is going to be described on clear linux documentation. Is necesary to
have a reference to the script for the documentation link. The script
will live in /usr/share/docs/ without afecting the performance or
functionality of tenserflow

Signed-off-by: vrodri3 <victor.rodriguez.bahena@intel.com>