Fixups for new tensorflow package
Because the content is all installed together in the ~/.local
directory, all but the first install needs --force-reinstall. Also
only do the final rename for the libtensorflow_framework.so after all
the mass avx512 and avx2 renames have happened to avoid weird renames
of .so ->.avx512 -> .so -> .avx2 -> .so.

Finally move the user directory installation content to the
buildroot.